Rename to VErrorSection and fix type
