+++ frontend/src/components/VErrorSection/VErrorSection.vue
